django_react_tyadmin/requirements.txt
Django~=4.2.7
djangorestframework~=3.14.0